black_box/requirements.txt
vllm>=0.6.0
transformers>=4.40.0
openai>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0